Cambios para primer deploy.

+++ src/app/movies/movie-details/movie-details.component.ts
@@ -40,7 +40,6 @@ export class MovieDetailsComponent implements OnInit {
   favoritesLabel = '';
   watchLaterLabel = '';
 
-
   // recordatorio hacer una tabla en la base de datos para que esto no sea estatico.
   blockedSites = ['divload', 'mangoplay', 'stream78'];
 
@@ -86,24 +85,31 @@ export class MovieDetailsComponent implements OnInit {
           if (this.movieLinks.size > 0) {
             this.movieOptionSelected = this.movieLinksKeys[0];
           }
-          this.watchLaterService.getByIdMovie(+url[1].path).subscribe( response => {
-            if (response['status'].includes('Success')) {
-              this.addedToWatchLaterList = true;
-              this.watchLaterLabel = 'Quitar De Lista';
-            } else {
-              this.addedToWatchLaterList = false;
-              this.watchLaterLabel = 'Ver Mas Tarde';
-            }
-          });
-          this.favoriteService.getByIdMovie(+url[1].path).subscribe( response => {
-            if (response['status'].includes('Success')) {
-              this.addedToFavoritesList = true;
-              this.favoritesLabel = 'Quitar De Favoritos';
-            } else {
-              this.addedToFavoritesList = false;
-              this.favoritesLabel = 'Agregar A Favoritos';
-            }
-          });
+          // Si se tiene el id del usuario es que si tiene la sesion iniciada
+          if (localStorage.getItem('id_user') != null) {
+            this.watchLaterService.getByIdMovieAndIdUser(+url[1].path, +localStorage.getItem('id_user')).subscribe( response => {
+              if (response['status'].includes('Success')) {
+                this.addedToWatchLaterList = true;
+                this.watchLaterLabel = 'Quitar De Lista';
+              } else {
+                this.addedToWatchLaterList = false;
+                this.watchLaterLabel = 'Ver Mas Tarde';
+              }
+            });
+            this.favoriteService.getByIdMovieAndIdUser(+url[1].path, +localStorage.getItem('id_user')).subscribe( response => {
+              if (response['status'].includes('Success')) {
+                this.addedToFavoritesList = true;
+                this.favoritesLabel = 'Quitar De Favoritos';
+              } else {
+                this.addedToFavoritesList = false;
+                this.favoritesLabel = 'Agregar A Favoritos';
+              }
+            });
+          } else {
+            // si llega aqui entonces no tiene la sesion iniciada
+            this.favoritesLabel = 'Agregar A Favoritos';
+            this.watchLaterLabel = 'Ver Mas Tarde';
+          }
         }
         this.commentSevice.getAllByIdMovie(+url[1].path).subscribe( comments => {
           if (comments['_embedded'] != null) {
@@ -137,6 +143,7 @@ export class MovieDetailsComponent implements OnInit {
                     }
                   });
                 }
+                this.contentLoaded = true;
               }, (failed: HttpErrorResponse) => {
                 this.snackbar.open(failed.message, '', {
                   duration: 3500, panelClass: ['error-snackbar']});
@@ -145,7 +152,6 @@ export class MovieDetailsComponent implements OnInit {
             this.snackbar.open(movie['message'], '', {
               duration: 3500, panelClass: ['error-snackbar']});
           }
-          this.contentLoaded = true;
         }, (error: HttpErrorResponse) => {
           this.snackbar.open(error.message, '', {
             duration: 3500, panelClass: ['error-snackbar']});
@@ -158,44 +164,56 @@ export class MovieDetailsComponent implements OnInit {
   }
 
   publishReview() {
-    const date = new Date();
-    this.review.idMovie = this.movie.idMovie;
-    this.review.publishedDate = date.toLocaleDateString();
-    this.review.publishedTime = date.toTimeString().split(' ')[0].substring(0, 5);
-    this.review.idReview = 0;
-    this.review.idUser = +localStorage.getItem('id_user');
-    this.review.username = localStorage.getItem('nombre');
-    this.reviewService.save(this.review).subscribe( response => {
-      this.snackbar.open(response['message']);
-      const newReview = new Review();
-      newReview.setReview(this.review);
-      newReview.idReview = response['idReview'];
-      this.reviews.push(newReview);
-      console.log(response);
-    }, (error: HttpErrorResponse) => {
-      this.snackbar.open(error.message, '', {
+    if (localStorage.getItem('id_user') != null) {
+      const date = new Date();
+      this.review.idMovie = this.movie.idMovie;
+      this.review.publishedDate = date.toLocaleDateString();
+      this.review.publishedTime = date.toTimeString().split(' ')[0].substring(0, 5);
+      this.review.idReview = 0;
+      this.review.idUser = +localStorage.getItem('id_user');
+      this.review.username = localStorage.getItem('nombre');
+      this.reviewService.save(this.review).subscribe( response => {
+        this.snackbar.open('Review Agregada.');
+        const newReview = new Review();
+        newReview.setReview(this.review);
+        newReview.idReview = response['idReview'];
+        this.reviews.push(newReview);
+        console.log(response);
+        this.review = new Review();
+      }, (error: HttpErrorResponse) => {
+          this.snackbar.open(error.message, '', {
+            duration: 3500, panelClass: ['error-snackbar']});
+      });
+    } else {
+      this.snackbar.open('Inicia sesion para poder publicar una review.', '', {
         duration: 3500, panelClass: ['error-snackbar']});
-    });
+    }
   }
 
   publishComment() {
-    const date = new Date();
-    this.comment.publishedTime = date.toTimeString().split(' ')[0].substring(0, 5);
-    this.comment.publishedDate = date.toLocaleDateString();
-    this.comment.idMovie = this.movie.idMovie;
-    this.comment.idUser = +localStorage.getItem('id_user');
-    this.comment.username = localStorage.getItem('nombre');
-    console.log(this.comment);
-    this.commentSevice.save(this.comment).subscribe( response => {
-      this.snackbar.open(response['message']);
-      const newComment = new Comment();
-      newComment.setComment(this.comment);
-      newComment.idComment = response['idComment'];
-      this.comments.push(newComment);
-    }, (error: HttpErrorResponse) => {
-      this.snackbar.open(error.message, '', {
+    if (localStorage.getItem('id_user') != null) {
+      const date = new Date();
+      this.comment.publishedTime = date.toTimeString().split(' ')[0].substring(0, 5);
+      this.comment.publishedDate = date.toLocaleDateString();
+      this.comment.idMovie = this.movie.idMovie;
+      this.comment.idUser = +localStorage.getItem('id_user');
+      this.comment.username = localStorage.getItem('nombre');
+      console.log(this.comment);
+      this.commentSevice.save(this.comment).subscribe( response => {
+        this.snackbar.open('Comentario Agregado.');
+        const newComment = new Comment();
+        newComment.setComment(this.comment);
+        newComment.idComment = response['idComment'];
+        this.comments.push(newComment);
+        this.comment = new Comment();
+      }, (error: HttpErrorResponse) => {
+        this.snackbar.open(error.message, '', {
+          duration: 3500, panelClass: ['error-snackbar']});
+      });
+    } else {
+      this.snackbar.open('Inicia sesion para poder dejar un comentario.', '', {
         duration: 3500, panelClass: ['error-snackbar']});
-    });
+    }
   }
 
   deleteComment(comment: Comment) {
@@ -268,44 +286,58 @@ export class MovieDetailsComponent implements OnInit {
   }
 
   modifyFavorites(movie: Movie) {
-    if (this.addedToFavoritesList) {
-      // eliminar
-      this.favoriteService.deleteByIdMovie(movie.idMovie).subscribe(response => {
-        if (response.status.includes('Success')) {
-          this.favoritesLabel = 'Agregar A Favoritos';
-          this.addedToFavoritesList = false;
-        }
-      });
+    if (localStorage.getItem('id_user') != null) {
+      if (this.addedToFavoritesList) {
+        // eliminar
+        this.favoriteService.deleteByIdMovie(movie.idMovie).subscribe(response => {
+          if (response.status.includes('Success')) {
+            this.favoritesLabel = 'Agregar A Favoritos';
+            this.addedToFavoritesList = false;
+            this.snackbar.open('Se elimino de la lista de favoritos.');
+          }
+        });
+      } else {
+        const fav = new Favorite();
+        fav.idFavorites = 0;
+        fav.idMovie = movie.idMovie;
+        fav.idUser = +localStorage.getItem('id_user');
+        this.favoriteService.save(fav).subscribe(response => {
+          this.favoritesLabel = 'Quitar De Favoritos';
+          this.addedToFavoritesList = true;
+          this.snackbar.open('Se agrego a la lista de favoritos.');
+        });
+      }
     } else {
-      const fav = new Favorite();
-      fav.idFavorites = 0;
-      fav.idMovie = movie.idMovie;
-      fav.idUser = +localStorage.getItem('id_user');
-      this.favoriteService.save(fav).subscribe(response => {
-        this.favoritesLabel = 'Quitar De Favoritos';
-        this.addedToFavoritesList = true;
-      });
+      this.snackbar.open('Inicia sesion para guardar en tu lista de favoritos.', '', {
+        duration: 3500, panelClass: ['error-snackbar']});
     }
   }
 
   modifyWatchLater(movie: Movie) {
-    if (this.addedToWatchLaterList) {
-      // eliminar
-      this.watchLaterService.deleteByIdMovie(movie.idMovie).subscribe(response => {
-        if (response.status.includes('Success')) {
-          this.watchLaterLabel = 'Ver Mas Tarde';
-          this.addedToWatchLaterList = false;
-        }
-      });
+    if (localStorage.getItem('id_user') != null) {
+      if (this.addedToWatchLaterList) {
+        // eliminar
+        this.watchLaterService.deleteByIdMovie(movie.idMovie).subscribe(response => {
+          if (response.status.includes('Success')) {
+            this.watchLaterLabel = 'Ver Mas Tarde';
+            this.addedToWatchLaterList = false;
+            this.snackbar.open('Se elimino de la lista de para ver mas tarde.');
+          }
+        });
+      } else {
+        const later = new WatchLater();
+        later.idWatchLater = 0;
+        later.idMovie = movie.idMovie;
+        later.idUser = +localStorage.getItem('id_user');
+        this.watchLaterService.save(later).subscribe( response => {
+          this.addedToWatchLaterList = true;
+          this.watchLaterLabel = 'Quitar De Lista';
+          this.snackbar.open('Se agrego a la lista para ver mas tarde.');
+        });
+      }
     } else {
-      const later = new WatchLater();
-      later.idWatchLater = 0;
-      later.idMovie = movie.idMovie;
-      later.idUser = +localStorage.getItem('id_user');
-      this.watchLaterService.save(later).subscribe( response => {
-        this.addedToWatchLaterList = true;
-        this.watchLaterLabel = 'Quitar De Lista';
-      });
+    this.snackbar.open('Inicia sesion para guardar en tu lista para ver despues.', '', {
+      duration: 3500, panelClass: ['error-snackbar']});
     }
   }
 }
